Changes to learner logic 1
