+++ platform-sdk/swirlds-virtualmap/src/main/java/com/swirlds/virtualmap/VirtualMapLearner.java
@@ -17,7 +17,6 @@ import com.swirlds.virtualmap.internal.hash.VirtualHasher;
 import com.swirlds.virtualmap.internal.merkle.VirtualMapMetadata;
 import com.swirlds.virtualmap.internal.merkle.VirtualMapStatistics;
 import com.swirlds.virtualmap.internal.reconnect.ConcurrentBlockingIterator;
-import com.swirlds.virtualmap.internal.reconnect.LearnerPullVirtualTreeView;
 import com.swirlds.virtualmap.internal.reconnect.ParallelSyncTraversalOrder;
 import com.swirlds.virtualmap.internal.reconnect.ReconnectHashLeafFlusher;
 import com.swirlds.virtualmap.internal.reconnect.ReconnectHashListener;
@@ -46,10 +45,9 @@ import org.hiero.consensus.reconnect.config.ReconnectConfig;
  * <p>Lifecycle:
  * <ul>
  *     <li>Constructor {@link #VirtualMapLearner(VirtualMap, ReconnectConfig, ReconnectMapStats)}</li>
- *     <li>Caller accesses {@link LearnerTreeView} via {@link #getLearnerView()} to start lessons using reconnect input/output streams.</li>
  *     <li>When synchronization starts, the teacher first sends its current leaf path range, and the {@link LearnerTreeView} implementation triggers {@link #init(long, long)}.</li>
  *     <li>Then on each dirty leaf {@link #onDirtyLeaf(VirtualLeafBytes)} has to be called</li>
- *     <li>On successful reconnect completion, the reconnect framework calls {@link #finish()} to finalize synchronization and create the new {@link VirtualMap} instance accessible via {@link #getVirtualMap()}.</li>
+ *     <li>On successful reconnect completion, the reconnect framework calls {@link #finish()} to finalize synchronization and return new {@link VirtualMap}.</li>
  *     <li>If reconnect fails before successful completion, the caller/reconnect orchestration code is responsible for aborting the reconnect attempt and cleaning up resources associated with the failed attempt via {@link #abortOnException()}.</li>
  * </ul>
  *
@@ -89,7 +87,6 @@ public final class VirtualMapLearner {
     private final VirtualDataSource dataSource;
     private final VirtualMapStatistics statistics;
     private final ReconnectHashLeafFlusher reconnectFlusher;
-    private final LearnerTreeView learnerView;
 
     private final VirtualMapMetadata reconnectState = new VirtualMapMetadata();
     private final ConcurrentBlockingIterator<VirtualLeafBytes> reconnectIterator =
@@ -109,13 +106,6 @@ public final class VirtualMapLearner {
     @Nullable
     private Hash finalHash;
 
-    /**
-     * The fully initialized {@link VirtualMap} created at the end of the reconnect process.
-     * Null until {@link #finish()} has been called.
-     */
-    @Nullable
-    private VirtualMap virtualMap;
-
     private enum Stage {
         NEW,
         INITIALIZING,
@@ -133,9 +123,7 @@ public final class VirtualMapLearner {
      * <p> Some original map fields are remembered so they can be used to construct new {@link VirtualMap} at the end of reconnect process.
      * The original map's data source is shut down and an independent copy is created for the reconnect process,
      * which will be updated with new leaves and hashes as they are received from the teacher.
-     *
-     * <p> {@link LearnerTreeView} will be eagerly created and available to access by {@link #getLearnerView()}.
-     *
+     **
      * @param originalMap the learner's current virtual map; must not be {@code null}
      * @param reconnectConfig reconnect configuration for this operation; must not be {@code null}
      * @param mapStats collector for reconnect metrics; must not be {@code null}
@@ -170,15 +158,6 @@ public final class VirtualMapLearner {
 
         reconnectFlusher =
                 new ReconnectHashLeafFlusher(dataSource, virtualMapConfig.reconnectFlushInterval(), statistics);
-        learnerView = buildLearnerView(reconnectConfig, mapStats);
-    }
-
-    /**
-     * @return {@link LearnerTreeView} that should be passed to the {@code LearningSynchronizer} to start the reconnect process.
-     */
-    @NonNull
-    public LearnerTreeView getLearnerView() {
-        return learnerView;
     }
 
     @NonNull
@@ -284,13 +263,13 @@ public final class VirtualMapLearner {
     /**
      * Signals that all nodes have been received from the teacher, then finalizes the reconnect
      * process. Waits for hashing to complete and creates the fully initialized {@link VirtualMap}.
-     * The new map can subsequently be retrieved via {@link #getVirtualMap()}.
      *
      * <p>This method is called automatically when the {@link LearnerTreeView} is closed.
      *
+     * @return synchronized virtual map
      * @throws MerkleSynchronizationException if hashing fails or if the calling thread is interrupted
      */
-    public void finish() {
+    public VirtualMap finish() {
         updateStage(Stage.INITIALIZED, Stage.FINISHING);
 
         logger.info(RECONNECT.getMarker(), "Finalizing learner reconnect");
@@ -299,11 +278,13 @@ public final class VirtualMapLearner {
         waitForHashingToComplete();
         reconnectFlusher.finish();
 
-        virtualMap = new VirtualMap(
+        VirtualMap virtualMap = new VirtualMap(
                 virtualMapConfig, dataSourceBuilder, dataSource, reconnectState.copy(), statistics, hasher, finalHash);
 
         updateStage(Stage.FINISHING, Stage.FINISHED);
         logger.info(RECONNECT.getMarker(), "Learner reconnect complete");
+
+        return virtualMap;
     }
 
     /**
@@ -503,13 +484,11 @@ public final class VirtualMapLearner {
      * <p>Must be called before passing this instance to a {@code LearningSynchronizer}.
      * The view will be closed automatically by the synchronizer when reconnect completes or fails.
      *
-     * @param reconnectConfig reconnect configuration
      * @param mapStats        collector for reconnect metrics
      * @return the learner tree view
      */
     @NonNull
-    private LearnerTreeView buildLearnerView(
-            @NonNull final ReconnectConfig reconnectConfig, @NonNull final ReconnectMapStats mapStats) {
+    private LearnerTreeView buildLearnerView(@NonNull final ReconnectMapStats mapStats) {
         logger.info(
                 RECONNECT.getMarker(),
                 "Building learner view for map with path range [{}, {}]",
@@ -518,12 +497,11 @@ public final class VirtualMapLearner {
 
         return switch (virtualMapConfig.reconnectMode()) {
             case VirtualMapReconnectMode.PULL_TOP_TO_BOTTOM ->
-                new LearnerPullVirtualTreeView(reconnectConfig, this, new TopToBottomTraversalOrder(), mapStats);
+                new LearnerTreeView(this, new TopToBottomTraversalOrder(), mapStats);
             case VirtualMapReconnectMode.PULL_TWO_PHASE_PESSIMISTIC ->
-                new LearnerPullVirtualTreeView(
-                        reconnectConfig, this, new TwoPhasePessimisticTraversalOrder(), mapStats);
+                new LearnerTreeView(this, new TwoPhasePessimisticTraversalOrder(), mapStats);
             case VirtualMapReconnectMode.PULL_PARALLEL_SYNC ->
-                new LearnerPullVirtualTreeView(reconnectConfig, this, new ParallelSyncTraversalOrder(), mapStats);
+                new LearnerTreeView(this, new ParallelSyncTraversalOrder(), mapStats);
             default ->
                 throw new UnsupportedOperationException("Unknown reconnect mode: "
                         + virtualMapConfig.reconnectMode()
@@ -533,24 +511,6 @@ public final class VirtualMapLearner {
     }
 
     // ---- Post-reconnect access ----
-
-    /**
-     * Returns the fully initialized {@link VirtualMap} created after reconnect completed.
-     *
-     * <p>Must only be called after the {@link LearnerTreeView} returned by
-     * {@link #getLearnerView()} has been closed.
-     *
-     * @return the reconnected, fully initialized virtual map
-     * @throws IllegalStateException if reconnect has not yet completed
-     */
-    @NonNull
-    public VirtualMap getVirtualMap() {
-        assert stage.get() == Stage.FINISHED : "Reconnect has not completed; current stage is " + stage.get();
-        if (virtualMap == null) {
-            throw new IllegalStateException("Reconnect has not completed - new virtual map is not yet available");
-        }
-        return virtualMap;
-    }
 
     /**
      * Destroy current reconnect state by closing iterator and datasource.
